+++ OgmentoAPI.Domain.Authorization.Abstraction/DataContext/UserRoles.cs
@@ -4,8 +4,8 @@ namespace OgmentoAPI.Domain.Authorization.Abstraction.DataContext
 {
     public partial class UserRoles
     {
-        public long UserId { get; set; }
-        public long RoleId { get; set; }
+        public int UserId { get; set; }
+        public int RoleId { get; set; }
 
         public virtual RolesMaster Role { get; set; }
         public virtual UsersMaster User { get; set; }
