+++ OgmentoAPI.Domain.Authorization.Abstraction/IAuthorizationContext.cs
@@ -7,7 +7,7 @@ namespace OgmentoAPI.Domain.Authorization.Abstraction
     {
         UsersMaster GetUserDetail(LoginModel login);
 
-        List<RolesMaster> GetUserRoles(long userID);
+        List<RolesMaster> GetUserRoles(int userID);
 
         
     }
